requirements.txt
# Web Framework ve Sunucu
fastapi==0.110.0
uvicorn==0.28.0

# Veritabanı (PostgreSQL) ve ORM
sqlalchemy==2.0.29
psycopg2-binary==2.9.9

# Veri Çekme (Web Scraping)
requests==2.31.0
beautifulsoup4==4.12.3

# Yapay Zeka (Gemini LLM)
google-generativeai==0.4.1

# Çevre Değişkenleri ve Veri Doğrulama
python-dotenv==1.0.1
pydantic==2.6.4